requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.1
Django==4.0
django-allauth==0.44.0
django-cors-headers==3.10.1
django-filter==21.1
django-simple-history==3.0.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.0
Markdown==3.3.6
oauthlib==3.1.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
zipp==3.6.0
gunicorn== 20.1.0
whitenoise==5.3.0
